+++ NEWS.md
@@ -4,6 +4,8 @@
 
 - Taking into account cuvette size to calculate the 1.5 threshold proposed by Kothawala when correcting for IFE.
 
+- Inner-filter effect correction factors are now corrected correctly. Because fluorescence is assumed to be measured in 1 cm cuvette, absorbance is now expressed per centimeter.
+
 # eemR 0.1.4
 
 - `eem_extract()` is now more intuitive to use. `remove` argument has been replace by `keep`. If `TRUE`, the specified samples will be returned. If `FALSE`, they will be removed (#37).